Update system-image-upgrader
This updates the upgrader code to first remove any file in the remove
list and only then apply the delta image.

That change will unfortunately cause the tarball to be unpacked twice,
though as it's limited to delta images, the impact should be fairly
minimal.

The problem this fixes is the case where a directory gets turned into a
symlink, in such case, tar will fail to overwrite the existing entry,
requiring it to be removed before unpacking.

Change-Id: Ia4d58e54494d5b419a0476b5d40b72b292df9dca
Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>
Signed-off-by: Ricardo Salveti de Araujo <ricardo.salveti@canonical.com>