system-image-upgrader: support ubuntu.img as the Ubuntu image name.
In order to stop the confusion around system.img from Android and
the system.img that contains the Ubuntu root filesystem start supporting
ubuntu.img as the name for the latter. Support the old naming too,
so that initramfs code that does not yet know about the change keeps working.

Change-Id: I6e35328104e9613cc86111029bec7edeb8aeecb7
Signed-off-by: Jani Monoses <jani.monoses@canonical.com>