Run e2fsck everytime an ext2/3/4 fs is mounted
The logic is split in two different places, the actual recovery binary
and the system-image-upgrader script, e2fsck is run prior to mounting.

Error checks have been added to verify the results for these to display
the broken upgrade image in recovery and to also better log a success or
a failure.

This also adds assurance that /data is mounted before calling the
system-image-upgrader script which depdends on it being mounted.

Change-Id: Ib4d7de2dc281a4bef2a3292f0baea0a100daed77
Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>